+++ collection.h
@@ -276,7 +276,7 @@ if (c != NULL) { c->freeCollection(c); }
  */
 #pragma pack(push, 4)
 typedef struct fiftyone_degrees_collection_header_t {
-	uint32_t startPosition; /**< Start position in the data file of the entities */
+	long startPosition; /**< Start position in the data file of the entities */
 	uint32_t length; /**< Length in bytes of all the entities */
 	uint32_t count; /**< Number of entities in the collection */
 } fiftyoneDegreesCollectionHeader;
